+++ src/packages/progress/progress.tsx
@@ -1,4 +1,4 @@
-import React, { FunctionComponent } from 'react'
+import React, { FunctionComponent, useEffect, useRef, useState } from 'react'
 import classNames from 'classnames'
 import { BasicComponent, ComponentDefaults } from '@/utils/typings'
 
@@ -9,6 +9,8 @@ export interface ProgressProps extends BasicComponent {
   strokeWidth: string
   showText: boolean
   animated: boolean
+  lazy: boolean
+  delay: number
 }
 
 const defaultProps = {
@@ -16,6 +18,8 @@ const defaultProps = {
   percent: 0,
   showText: false,
   animated: false,
+  lazy: false,
+  delay: 0,
 } as ProgressProps
 
 export const Progress: FunctionComponent<
@@ -31,6 +35,8 @@ export const Progress: FunctionComponent<
     showText,
     animated,
     children,
+    lazy,
+    delay,
     ...rest
   } = {
     ...defaultProps,
@@ -49,19 +55,83 @@ export const Progress: FunctionComponent<
     background,
   }
 
+  const [displayPercent, setDispalyPercent] = useState(0)
+
   const stylesInner: React.CSSProperties = {
-    width: `${percent}%`,
+    width: `${displayPercent}%`,
     background: color,
   }
 
+  useEffect(() => {
+    setDispalyPercent(percent)
+  }, [percent])
+
+  const [intersecting, setIntersecting] = useState(false)
+
+  const progressRef = useRef(null)
+  const observer: any = useRef(null)
+  const initObserver = () => {
+    const options = {
+      threshold: [0],
+      rootMargin: '0px',
+    }
+    observer.current = new IntersectionObserver((entires, self) => {
+      entires.forEach((item) => {
+        if (item.isIntersecting) {
+          setIntersecting(true)
+        } else {
+          setIntersecting(false)
+        }
+      })
+    }, options)
+    observer.current.observe(progressRef.current)
+  }
+
+  const resetObserver = () => {
+    observer.current.disconnect && observer.current.disconnect()
+  }
+
+  useEffect(() => {
+    if (lazy) {
+      setTimeout(() => {
+        if (intersecting) {
+          setDispalyPercent(percent)
+        } else {
+          setDispalyPercent(0)
+        }
+      }, delay)
+    }
+  }, [intersecting])
+
+  useEffect(() => {
+    lazy && initObserver()
+    let timer: any = null
+    if (delay) {
+      setDispalyPercent(0)
+      timer = setTimeout(() => {
+        setDispalyPercent(percent)
+      }, delay)
+    }
+
+    return () => {
+      lazy && resetObserver()
+      timer && clearTimeout(timer)
+    }
+  }, [])
+
   return (
-    <div className={classNames(classPrefix, className)} style={style} {...rest}>
+    <div
+      ref={progressRef}
+      className={classNames(classPrefix, className)}
+      style={style}
+      {...rest}
+    >
       <div className={`${classPrefix}-outer`} style={stylesOuter}>
         <div className={classesInner} style={stylesInner}>
           {showText && (
             <div
               className={`${classPrefix}-text`}
-              style={{ left: `${percent}%` }}
+              style={{ left: `${displayPercent}%` }}
             >
               {children || (
                 <div