+++ src/packages/swiper/swiper.taro.tsx
@@ -2,6 +2,7 @@ import React, {
   Children,
   forwardRef,
   ReactNode,
+  useEffect,
   useImperativeHandle,
   useRef,
   useState,
@@ -18,7 +19,7 @@ import { Indicator } from '../indicator/indicator.taro'
 
 export interface SwiperProps extends Omit<TaroSwiperProps, 'ref'> {
   width: number
-  height: number
+  height: number | string
   direction: 'horizontal' | 'vertical'
   indicator: ReactNode
   autoPlay: boolean
@@ -54,14 +55,17 @@ export const Swiper = forwardRef((props: Partial<SwiperProps>, ref) => {
   }
   const [current, setCurrent] = useState(defaultValue)
   const childrenCount = useRef(Children.toArray(children).length)
+  useEffect(() => {
+    setCurrent(defaultValue)
+  }, [defaultValue])
   const renderIndicator = () => {
     if (React.isValidElement(indicator)) return indicator
     if (indicator === true) {
       return (
         <div
           className={classNames({
-            [`${classPrefix}__pagination`]: true,
-            [`${classPrefix}__pagination-vertical`]: direction === 'vertical',
+            [`${classPrefix}__indicator`]: true,
+            [`${classPrefix}__indicator-vertical`]: direction === 'vertical',
           })}
         >
           <Indicator