+++ src/packages/swiper/swiper.tsx
@@ -43,7 +43,7 @@ const defaultProps = {
   autoPlay: 0,
   direction: 'horizontal',
   indicator: false,
-  loop: true,
+  loop: false,
   touchable: true,
   preventDefault: true,
   stopPropagation: true,
@@ -166,10 +166,12 @@ export const Swiper = React.forwardRef<
     resetPosition()
     resetTouchDetails()
     requestFrame(() => {
-      swiperRef.current.moving = false
-      move({
-        pace: -1,
-        isEmit: true,
+      requestFrame(() => {
+        swiperRef.current.moving = false
+        move({
+          pace: -1,
+          isEmit: true,
+        })
       })
     })
   }
@@ -178,10 +180,12 @@ export const Swiper = React.forwardRef<
     resetPosition()
     resetTouchDetails()
     requestFrame(() => {
-      swiperRef.current.moving = false
-      move({
-        pace: 1,
-        isEmit: true,
+      requestFrame(() => {
+        swiperRef.current.moving = false
+        move({
+          pace: 1,
+          isEmit: true,
+        })
       })
     })
   }
@@ -190,16 +194,18 @@ export const Swiper = React.forwardRef<
     resetPosition()
     resetTouchDetails()
     requestFrame(() => {
-      swiperRef.current.moving = false
-      let targetIndex
-      if (props.loop && childCount === index) {
-        targetIndex = active === 0 ? 0 : index
-      } else {
-        targetIndex = index % childCount
-      }
-      move({
-        pace: targetIndex - active,
-        isEmit: true,
+      requestFrame(() => {
+        swiperRef.current.moving = false
+        let targetIndex
+        if (props.loop && childCount === index) {
+          targetIndex = active === 0 ? 0 : index
+        } else {
+          targetIndex = index % childCount
+        }
+        move({
+          pace: targetIndex - active,
+          isEmit: true,
+        })
       })
     })
   }
@@ -400,7 +406,7 @@ export const Swiper = React.forwardRef<
   }
 
   useEffect(() => {
-    swiperRef.current.activePagination = (active + childCount) % childCount
+    swiperRef.current.activeIndicator = (active + childCount) % childCount
   }, [active])
 
   const init = (active: number = +propSwiper.defaultValue) => {
@@ -486,8 +492,8 @@ export const Swiper = React.forwardRef<
       return (
         <div
           className={classNames({
-            [`${classPrefix}__pagination`]: true,
-            [`${classPrefix}__pagination-vertical`]: isVertical,
+            [`${classPrefix}__indicator`]: true,
+            [`${classPrefix}__indicator-vertical`]: isVertical,
           })}
         >
           <Indicator